requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
requests==2.31.0
pypdf==3.17.0
python-docx==1.1.0
docx2txt==0.8
groq==0.4.1
faiss-cpu==1.7.4  # Alternative to chromadb that doesn't need Rust
numpy==1.24.3
scikit-learn==1.3.0